+++ examples/prototype/AirlockPowerdown.tq
@@ -1,6 +1,6 @@
 % technique v1
 ! PD; © 2003 National Aeronautics and Space Administration, Canadian Space Agency, European Space Agency, and Others
-& nasa-flight-plan,v4.0
+& nasa-esa-iss,v4.0
 
 emergency_procedures :
