+++ src/main.rs
@@ -9,7 +9,7 @@ use tracing_subscriber::{self, EnvFilter};
 use technique::formatting::{self, Identity};
 use technique::highlighting::{self, Terminal};
 use technique::parsing;
-use technique::templating::{self, Checklist, NasaFlightPlan, Procedure, Recipe, Source};
+use technique::templating::{self, Checklist, NasaEsaIss, Procedure, Recipe, Source};
 
 mod editor;
 mod output;
@@ -124,7 +124,7 @@ fn main() {
                     Arg::new("domain")
                         .short('d')
                         .long("domain")
-                        .value_parser(["checklist", "nasa-flight-plan", "procedure", "recipe", "source"])
+                        .value_parser(["checklist", "nasa-esa-iss", "procedure", "recipe", "source"])
                         .action(ArgAction::Set)
                         .help("The kind of procedure this Technique document represents. By default the value specified in the input document's metadata will be used, falling back to source if unspecified."),
                 )
@@ -342,7 +342,7 @@ fn main() {
                     .unwrap_or("source"),
             };
 
-            // Strip version suffix (e.g. "nasa-flight-plan,v4.0" -> "nasa-flight-plan")
+            // Strip version suffix (e.g. "nasa-esa-iss,v4.0" -> "nasa-esa-iss")
             let domain = domain
                 .split(',')
                 .next()
@@ -354,7 +354,7 @@ fn main() {
             let template: &dyn templating::Template = match domain {
                 "source" => &Source,
                 "checklist" => &Checklist,
-                "nasa" | "nasa-flight-plan" => &NasaFlightPlan,
+                "nasa-esa-iss" => &NasaEsaIss,
                 "procedure" => &Procedure,
                 "recipe" => &Recipe,
                 other => {
